Did some work

+++ SENG201/src/DiceRoll.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 public class DiceRoll implements VillainGame {
 	
 	private static final int MAX_VAL = 1000;
-	private static final double CHANCE_VAL = 700;
+	private static final int CHANCE_VAL = 700;
 	
 	@Override
 	public boolean play(String villain_name, Hero playing_hero) {
@@ -20,7 +20,7 @@ public class DiceRoll implements VillainGame {
 		int player_score = 0;
 		int hero_roll = 0;
 		int villain_roll = 0;
-		double illusion = playing_hero.getIllusion();
+		int illusion = playing_hero.getIllusion();
 		
 		m.displayMessage("This is a best of three games of roll the dice, roll a higher number to win");
 		
@@ -55,7 +55,7 @@ public class DiceRoll implements VillainGame {
 		}
 		
 		int ran_chance = num.nextInt(MAX_VAL);
-		int win_chance = (int) (CHANCE_VAL / illusion);
+		int win_chance = CHANCE_VAL / illusion;
 		
 		if (player_score == 2) {
 			m.displayMessage("Congratualtions. " + playing_hero.getName() + " beat " + villain_name);
